wip: make electron tests run properly
