+++ package.json
@@ -11,6 +11,7 @@
     "@typescript-eslint/parser": "^8.56.1",
     "husky": "^9.1.7",
     "lint-staged": "^16.3.1",
+    "npm-run-all": "^4.1.5",
     "prettier": "^3.8.1",
     "tsx": "catalog:",
     "typescript": "catalog:",
