+++ packages/app-desktop/package.json
@@ -40,6 +40,7 @@
     "@types/fs-extra": "^11.0.4",
     "@types/lodash": "^4.17.20",
     "@types/semver": "^7.7.1",
+    "better-sqlite3": "^12.6.2",
     "electron-log": "^5.4.3",
     "extract-zip": "^2.0.1",
     "font-list": "^2.0.1",
