+++ packages/app-desktop/src/note/note.dao.ts
@@ -1,9 +1,7 @@
+import { NotFoundError, ParentId, Rank } from "@darkwrite/common";
 import { EntityManager, IsNull, Like, Repository } from "typeorm";
 import { AppDataSource } from "../db";
 import { Note } from "../entity";
-import { ParentId } from "@darkwrite/common";
-import { Rank } from "@darkwrite/common";
-import { NotFoundError } from "@darkwrite/common";
 
 export class NoteDAO {
   constructor(
