+++ packages/common/src/index.ts
@@ -6,17 +6,13 @@ export * from "./dto/request/embed.request";
 export * from "./dto/request/note.request";
 export * from "./dto/request/workspace.request";
 export * from "./dto/index";
-export * from "./rank-correction.test";
 export * from "./json-util";
 export * from "./rank";
 export * from "./note-content";
 export * from "./note";
 export * from "./contract";
 export * from "./rank-correction";
-export * from "./sanitizer.test";
-export * from "./note.test";
 export * from "./settings";
-export * from "./markdown.test";
 export * from "./markdown";
 export * from "./settings-migrator";
 export * from "./note-filters";
@@ -27,7 +23,6 @@ export * from "./note-customization";
 export * from "./os";
 export * from "./client";
 export * from "./object";
-export * from "./rank.test";
 export * from "./theme";
 export * from "./sanitizer";
 export * from "./id";
@@ -35,14 +30,11 @@ export * from "./ts-util";
 export * from "./error/client-error";
 export * from "./error/index";
 export * from "./error/common-errors";
-export * from "./theme.test";
 export * from "./settings-v1-schema";
 export * from "./themes/default-themes";
 export * from "./themes/catppuccin";
 export * from "./themes/darkwrite-black";
 export * from "./themes/index";
 export * from "./themes/darkwrite-default";
-export * from "./event/bus.test";
 export * from "./event/bus";
 export * from "./field";
-export * from "./object.test";
