+++ pnpm-workspace.yaml
@@ -21,3 +21,4 @@ onlyBuiltDependencies:
   - electron
   - electron-winstaller
   - esbuild
+  - better-sqlite3
